create mockMoviesData.js
