+++ src/components/MovieCard.js
@@ -1,4 +1,5 @@
-import React from 'react'
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 export default function MovieCard({ title, styles }) {
     return (
