+++ src/services/mockMoviesData.js
@@ -1,0 +1,34 @@
+const movies = [
+    {
+        title: 'Movie 1'
+    },
+    {
+        title: 'Movie 2'
+    },
+    {
+        title: 'Movie 3'
+    },
+    {
+        title: 'Movie 4'
+    },
+    {
+        title: 'Movie 5'
+    },
+    {
+        title: 'Movie 6'
+    },
+    {
+        title: 'Movie 7'
+    },
+    {
+        title: 'Movie 8'
+    },
+    {
+        title: 'Movie 9'
+    },
+    {
+        title: 'Movie 10'
+    }
+];
+
+export default movies;
